Organized the code into directories
