+++ include/Algorithms/Vanama.h
@@ -9,16 +9,7 @@
 #include <set>
 #include <algorithm>
 #include <random>
-#include "Graph/graph.h"
-#include "GreedySearch.h"
-
-// Type Alias
-typedef GraphNode<DataVector<float>> GreedySearchNode;
-typedef Graph<DataVector<float>> GreedySearchGraph;
-typedef std::set<GreedySearchNode> GreedySearchNodeSet;
-#include <algorithm>
-#include <random>
-#include "Graph/graph.h"
+#include "../DataStructures/Graph/graph.h"
 #include "GreedySearch.h"
 
 // Type Alias
