--- knn.cpp
@@ -1,12 +1,0 @@
-// CPP libraries
-// this contains most of the STL headers files (data structures) amd iostream
-// WARNING individual #inlcludes to ensure portability
-// temporary placeholder for clean code
-#include <bits/stdc++.h>
-
-// Custom header files
-#include "GreedySearch.h"
-#include "distance.h"
-#include "vanama.h"
-
-//this should appear in main branch
