--- src/vamana_creation.cpp
@@ -1,23 +1,0 @@
-//######################################################################################
-// This file includes:
-// ~> Create_Vamana_Index
-//######################################################################################
-
-// CPP libraries
-// this contains most of the STL headers files (data structures) amd iostream
-// WARNING individual #inlcludes to ensure portability
-// temporary placeholder for clean code
-#include <bits/stdc++.h>
-
-// Custom header files
-#include "GreedySearch.h"
-#include "distance.h"
-#include "vanama.h"
-
-//######################################################################################
-
-// DEFINITIONS
-
-void Create_Vamana_Index(void){
-
-}
