+++ tests/test_recall.cc
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <set>
-#include "../include/recall.h"
+#include "../include/Evaluation/recall.h"
 #include "../include/acutest.h"
